New Feature:    Added light cones to light. Useful for debugging lights or for    providing a visual cue.
Fixes:
    Rod tesselators renders disc with correct orientation when AddTube method is used

API:
RotateAboutX, RotateAboutY, RotateAboutZ and MakeTranslateMatrix
added to Sys.Geometry.F32

+++ hyperverse/hv.app.cpp
@@ -43,7 +43,7 @@ namespace HV
 			return e.platform.scene.GetSkyboxCubeId();
 		}
 
-		const Light* GetLights(size_t& nLights) const override
+		const Light* GetLights(uint32& nLights) const override
 		{
 			return e.platform.scene.GetLights(nLights);
 		}
